+++ core/renderer.py
@@ -13,7 +13,6 @@ import time
 
 glfw.init()
 
-STEP = 512
 VERTICES_SIZE = 256*16*16*24*3
 TEXCOORDS_SIZE = 256*16*16*24*2
 
@@ -156,34 +155,3 @@ class TerrainRenderer:
                 self.fps = 1 / (self.timings[-1] - self.timings[0])
             except:
                 self.fps = 0
-
-class TerrainMeshStorage:
-    def __init__(self):
-        self.vertices = []
-        self.texCoords = []
-    
-    def add(self, posList, texCoords):
-        self.vertices.append(posList)
-        self.texCoords.append(texCoords)
-
-    def clear(self):
-        self.vertices = []
-        self.texCoords = []
-
-    def _group(self):
-        to_add = []
-        for i in range(0, len(self.vertices), STEP):
-            verts = self.vertices[i:i+STEP]
-            tex = self.texCoords[i:i+STEP]
-
-            _verts = []
-            _tex = []
-
-            for i in verts:
-                _verts.extend(i)
-            for i in tex:
-                _tex.extend(i)
-
-            to_add.append((_verts, _tex))
-
-        return to_add
