--- helpers/biomes/__init__.py
@@ -1,5 +1,0 @@
-from helpers.biomes.plains import *
-from helpers.biomes.desert import *
-
-plains = PlainsGenerator()
-desert = DesertGenerator()
